refactor(backend): Implement getUserById where necessary
