+++ backend/src/middleware/authMiddleware.ts
@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from "express";
 
-import { findUserById } from "../controllers/findOrCreateUser.js";
 import { verifyJWT } from "../utils/verifyJWT.js";
+import { getUserById } from "../controllers/userController.js";
 
 export const authMiddleware = async (
   req: Request,
@@ -31,9 +31,9 @@ export const authMiddleware = async (
     }
 
     // Find user - decoded.sub is user id
-    const user = await findUserById(String(verifiedToken.sub));
+    const user = await getUserById(String(verifiedToken.sub));
     if (!user) {
-      res.status(401).json({ error: "User not found" });
+      res.status(404).json({ error: "User not found" });
       return;
     }
